Remove repeated code
